# Minor performance improvement for DfaMatcherBuilder