* Decompose list of policies in separate array without using **Linq**.
This avoids iterating over all policies in `AddNode(DfaNode node, DfaState[] states, int exitDestination)`, and improves performance of `.ctor`, `Build()` and `BuildDfaTree(bool includeLabel = false)`.
* When endpoint is not a **RouteEndpoint**, do not bother clearing lists in `CreateCandidate(Endpoint endpoint, int score)` and use empty arrays to reduce allocations.

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17134
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3914069 Hz, Resolution=255.4886 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host]     : .NET Core 2.2.0-preview3-26927-02 (CoreCLR 4.6.26927.02, CoreFX 4.6.26927.02), 64bit RyuJIT
  Job-VYAOMV : .NET Core 2.2.0-preview3-26927-02 (CoreCLR 4.6.26927.02, CoreFX 4.6.26927.02), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.2  
RunStrategy=Throughput  

```

**Before**

|            Method |      Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|------------------ |----------:|----------:|----------:|----------:|-------:|----------:|
|              Ctor |  2.123 us | 0.0102 us | 0.0095 us | 470,998.0 | 0.0191 |   1.81 KB |
| ConstructAndBuild | 14.016 us | 0.0670 us | 0.0594 us |  71,347.4 | 0.1526 |   13.5 KB |

**After**

|            Method |      Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|------------------ |----------:|----------:|----------:|----------:|-------:|----------:|
|              Ctor |  1.265 us | 0.0045 us | 0.0038 us | 790,604.1 | 0.0210 |    1.8 KB |
| ConstructAndBuild | 12.562 us | 0.1081 us | 0.0959 us |  79,604.4 | 0.1526 |  13.49 KB |