@rynowak 2018-10-12
The changes here seem straightforward and positive. What's the impact on https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherBuilderAzureBenchmark.cs