@drieseng 2018-10-12
I was planning to measure the impact on that benchmark once I finished submitting a few more PRs.
I don't expect one of these PRs individually to cause a ripple, but combined they have a more noticeable effect. Is it ok to update this PR with the combined effect of these PRs?