@rynowak 2018-10-12
> Is it ok to update this PR with the combined effect of these PRs?
Sure

I think the set of benchmarks we care about for the builder is:
[Large] (https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherBuilderAzureBenchmark.cs)
[Medium] (https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherBuilderGithubBenchmark.cs)
[Small] (https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherBuilderMultipleEntryBenchmark.cs)

If we're missing scenarios, I'd be happy to see them added.

---

I also don't want to discourage you from adding value, but to be clear - this is the code path used once at startup. These types play no role in processing requests.